requirements.txt
fastapi
uvicorn[standard]
beanie
motor
pydantic[email]
python-dotenv
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
python-jose[cryptography]